Moved util function to correct scope
